LPS-36216 Replace getXYByUuidAndGroupId methods to fetchXYByUuidAndGroupId now that we have them for all models

+++ portal-impl/src/com/liferay/portlet/documentlibrary/lar/FolderStagedModelDataHandler.java
@@ -65,10 +65,13 @@ public class FolderStagedModelDataHandler
 			String uuid, long groupId, String className, String extraData)
 		throws PortalException, SystemException {
 
-		DLFolder folder = DLFolderLocalServiceUtil.getDLFolderByUuidAndGroupId(
-			uuid, groupId);
+		DLFolder folder =
+			DLFolderLocalServiceUtil.fetchDLFolderByUuidAndGroupId(
+				uuid, groupId);
 
-		DLFolderLocalServiceUtil.deleteFolder(folder);
+		if (folder != null) {
+			DLFolderLocalServiceUtil.deleteFolder(folder);
+		}
 	}
 
 	@Override
